hitterRank bug fix

+++ src/main/java/com/demo/project/crawling/repository/hitterRankRepo.java
@@ -3,5 +3,5 @@ package com.demo.project.crawling.repository;
 import com.demo.project.crawling.model.hitterRank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface hitterRankRepo extends JpaRepository<hitterRank, Integer> {
+public interface hitterRankRepo extends JpaRepository<hitterRank, Long> {
 }
